cpp: bioformats: Move TileBuffer out from tiff namespace
This is also usable in other readers and writers.  Remove
use of TIFF-specific memory functions in place of standard
C++ array allocation.
