+++ cpp/lib/ome/bioformats/CMakeLists.txt
@@ -48,6 +48,7 @@ set(OME_BIOFORMATS_SOURCES
     Modulo.cpp
     PixelBuffer.cpp
     PixelProperties.cpp
+    TileBuffer.cpp
     UnknownFormatException.cpp
     UnsupportedCompressionException.cpp
     VariantPixelBuffer.cpp
@@ -69,6 +70,7 @@ set(OME_BIOFORMATS_HEADERS
     Modulo.h
     PixelBuffer.h
     PixelProperties.h
+    TileBuffer.h
     Types.h
     UnknownFormatException.h
     UnsupportedCompressionException.h
@@ -100,7 +102,6 @@ set(OME_BIOFORMATS_TIFF_SOURCES
     tiff/Sentry.cpp
     tiff/Tags.cpp
     tiff/TIFF.cpp
-    tiff/TileBuffer.cpp
     tiff/TileCache.cpp
     tiff/TileCoverage.cpp
     tiff/TileInfo.cpp)
@@ -113,7 +114,6 @@ set(OME_BIOFORMATS_TIFF_HEADERS
     tiff/Sentry.h
     tiff/Tags.h
     tiff/TIFF.h
-    tiff/TileBuffer.h
     tiff/TileCache.h
     tiff/TileCoverage.h
     tiff/TileInfo.h
