+++ cpp/lib/ome/bioformats/tiff/IFD.cpp
@@ -43,13 +43,13 @@
 #include <ome/internal/config.h>
 
 #include <ome/bioformats/PlaneRegion.h>
+#include <ome/bioformats/TileBuffer.h>
 #include <ome/bioformats/tiff/IFD.h>
 #include <ome/bioformats/tiff/Tags.h>
 #include <ome/bioformats/tiff/Field.h>
 #include <ome/bioformats/tiff/TIFF.h>
 #include <ome/bioformats/tiff/Sentry.h>
 #include <ome/bioformats/tiff/Exception.h>
-#include <ome/bioformats/tiff/TileBuffer.h>
 #include <ome/bioformats/tiff/TileCache.h>
 
 #include <ome/compat/thread.h>
