+++ cpp/test/ome-bioformats/planeregion.cpp
